Move BasicBuild to new integration test framework and implement
