--- src/VisualStudio/IntegrationTest/IntegrationTests/VisualBasic/BasicBuild.cs
@@ -1,50 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-#nullable disable
-
-using System.Threading.Tasks;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.VisualStudio.IntegrationTest.Utilities;
-using Roslyn.Test.Utilities;
-using Xunit;
-using Xunit.Abstractions;
-using ProjectUtils = Microsoft.VisualStudio.IntegrationTest.Utilities.Common.ProjectUtils;
-
-namespace Roslyn.VisualStudio.IntegrationTests.VisualBasic
-{
-    [Collection(nameof(SharedIntegrationHostFixture))]
-    public class BasicBuild : AbstractIntegrationTest
-    {
-        public BasicBuild(VisualStudioInstanceFactory instanceFactory)
-            : base(instanceFactory)
-        {
-        }
-
-        public override async Task InitializeAsync()
-        {
-            await base.InitializeAsync().ConfigureAwait(true);
-            VisualStudio.SolutionExplorer.CreateSolution(nameof(BasicBuild));
-            var testProj = new ProjectUtils.Project("TestProj");
-            VisualStudio.SolutionExplorer.AddProject(testProj, WellKnownProjectTemplates.ConsoleApplication, LanguageNames.VisualBasic);
-        }
-
-        [WpfFact, Trait(Traits.Feature, Traits.Features.Build)]
-        public void BuildProject()
-        {
-            var editorText = @"Module Program
-
-    Sub Main()
-        Console.WriteLine(""Hello, World!"")
-    End Sub
-
-End Module";
-
-            VisualStudio.Editor.SetText(editorText);
-
-            // TODO: Validate build works as expected
-        }
-    }
-}
